feat: update tests for bank forms
